+++ ft_memcpy.c
@@ -6,15 +6,17 @@
 /*   By: jehelee <jehelee@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/10 16:55:52 by jehelee           #+#    #+#             */
-/*   Updated: 2022/11/10 18:37:59 by jehelee          ###   ########.fr       */
+/*   Updated: 2022/11/15 12:38:58 by jehelee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dst, const void *src, unsigned long len)
+#include "libft.h"
+
+void	*ft_memcpy(void *dst, const void *src, size_t len)
 {
-	char			*dst_tmp;
-	const char		*src_tmp;
-	unsigned long	i;
+	char		*dst_tmp;
+	const char	*src_tmp;
+	size_t		i;
 
 	dst_tmp = (char *) dst;
 	src_tmp = (const char *) src;
